Simplify outbound request sink structure
